Fix logic error in back_copy
In general, when there are multiple overlapping views then the trick of using
beta=0 on the first round of back-copying and beta=1 on subsequent iterations
does not get the right answer.

The new code is more conservative, doing a first gemv call to simply clear all
outputs (possibly writing 0 to some output locations multiple times) and then
using beta=1 gemvs to incrementally load the new values into the output areas.